refactor(ScreeningListViewItemUIState): sealed class의 자식 클래스들을 다른 파일로 분리
