+++ app/src/main/java/woowacourse/movie/view/activities/home/fragments/screeninglist/AdvertisementUIState.kt
@@ -1,0 +1,6 @@
+package woowacourse.movie.view.activities.home.fragments.screeninglist
+
+import androidx.annotation.DrawableRes
+
+data class AdvertisementUIState(@DrawableRes val advertisementImage: Int) :
+    ScreeningListViewItemUIState()
